+++ src/cli/mod.rs
@@ -118,6 +118,7 @@ pub enum KeywordColor {
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Hash)]
 pub enum Extra {
     Ipv6,
+    JvmStackTrace,
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Hash)]